move redundant code to separate files
