+++ src/specimens/ReactSpecimen/ReactSpecimen.js
@@ -4,9 +4,11 @@ import Radium from 'radium';
 import Frame from '../../components/Frame/Frame';
 import Specimen from '../../components/Specimen/Specimen';
 import HighlightedCode from '../../components/HighlightedCode/HighlightedCode';
+import ResponsiveTabs from '../../components/ResponsiveTabs/ResponsiveTabs';
 import Hint from '../../specimens/Hint';
 import reactElementToString from './reactElementToString';
 import transformJSX from '../../utils/transformJSX';
+import validateSizes from '../../utils/validateSizes';
 
 function getStyle(theme) {
   return {
@@ -52,96 +54,9 @@ function getStyle(theme) {
       overflow: 'hidden',
       padding: '15px',
       textAlign: 'center'
-    },
-    tabContainer: {
-      alignItems: 'center',
-      background: 'white',
-      borderBottom: `1px solid #eee`,
-      display: 'flex',
-      overflowX: 'scroll',
-      width: '100%'
-    },
-    tab: {
-      boxSizing: 'border-box',
-      alignItems: 'center',
-      background: '#eee',
-      color: theme.bgDark,
-      cursor: 'pointer',
-      display: 'flex',
-      flexGrow: 1,
-      opacity: 0.3,
-      padding: '12px',
-      transitionDuration: '.2s, .4s',
-      transitionProperty: 'width, background-color, border, opacity',
-      ':hover': {
-        color: 'red'
-      }
-    },
-    tabActive: {
-      background: 'white',
-      color: theme.textColor,
-      cursor: 'auto',
-      opacity: 1
-    },
-    tabDimension: {
-      color: theme.bgDark,
-      display: 'inline',
-      fontFamily: theme.fontMono,
-      marginLeft: 5,
-      opacity: 0.3
     }
   };
 }
-
-/**
- * Generates a small preview to show the aspect ratio
- */
-const Preview = ({proportion}) => {
-  if ( !proportion ) null;  return (
-    <div style={{width: '30px', height: '30px', display: 'inline-block', marginRight: 5}}>
-      <svg viewBox={`0 0 2 2`}>
-        <rect style={{fill: '#ccc'}} width={proportion} height={1} x={(2 - proportion) * 0.5} y='0.5'/>
-      </svg>
-    </div>);
-};
-
-
-/**
- * Checks if the delivered props are valid, returns false if not, otherwise a filtered Array
- */
-const validateDevices = (input, catalogDevices) => {
-  const isArray = Array.isArray(input);
-  if (input === true) {
-    return catalogDevices;
-  } else if (typeof input === 'string') {
-    const foundInList = catalogDevices.find(val => input === val.name);
-    return foundInList ? [].concat(foundInList) : false;
-  } else if (isArray && input.length === input.filter((item=> typeof item === 'string')).length) {
-    const filtered = catalogDevices.filter( val => {
-      return input.find( d=> d === val.name);
-    });
-    return filtered.length === input.length ? filtered : false;
-  }
-  return false;
-};
-
-
-const ResponsiveTabs = ( {deviceList, action, activeDevice, styles, parentWidth} ) => {
-  return (<div style={styles.tabContainer}>
-    {deviceList.map((val, i)=>{
-      const isTabActive = activeDevice.name === val.name;
-      const activeStyles = isTabActive && styles.tabActive;
-      return (<div key={i} style={{...styles.tab, ...activeStyles}} onClick={action.bind(this, val)}>
-        <Preview proportion={val.width / val.height}/>
-        {val.name}{isTabActive &&
-          <div style={styles.tabDimension}>
-            {activeDevice.width}×{activeDevice.height}
-            {isTabActive && parentWidth <= val.width && ', scaled'}
-          </div>}
-      </div>);
-    })}
-  </div>);
-};
 
 class ReactSpecimen extends Component {
   constructor(props) {
@@ -149,7 +64,7 @@ class ReactSpecimen extends Component {
     this.state = {
       elementState: props.state,
       parentWidth: null,
-      screenSize: validateDevices(props.responsive, props.catalog.devices)[0] || null
+      screenSize: validateSizes(props.responsive, props.catalog.devices)[0] || null
     };
     this.setElementState = this.setElementState.bind(this);
     this.setSize = this.setSize.bind(this);
@@ -192,7 +107,7 @@ class ReactSpecimen extends Component {
     const {catalog: {page: {imports}, theme, devices}, children, noSource, frame, ...options} = this.props;
     const {screenSize: device, parentWidth} = this.state;
     const styles = getStyle(theme);
-    const validDevices = validateDevices(options.responsive, devices);
+    const validSizes = validateSizes(options.responsive, devices);
 
     const exampleStyles = {
       ...(options.plain ? styles.plain : null),
@@ -222,7 +137,7 @@ class ReactSpecimen extends Component {
       code = reactElementToString(children);
     }
 
-    if (options.responsive && !validDevices) {
+    if (options.responsive && !validSizes) {
       return <Hint warning>Please check that the responsive parameters match an existing entry.</Hint>;
     }
 
@@ -231,7 +146,7 @@ class ReactSpecimen extends Component {
     return (
       <section style={styles.container} ref='specimen'>
         {device &&
-          <ResponsiveTabs styles={styles} deviceList={validDevices} action={this.setSize} activeDevice={device} parentWidth={parentWidth}/>
+          <ResponsiveTabs theme={theme} deviceList={validSizes} action={this.setSize} activeDevice={device} parentWidth={parentWidth}/>
         }
         <div style={{...styles.content, ...exampleStyles}}>
           {frame || device
@@ -246,16 +161,6 @@ class ReactSpecimen extends Component {
   }
 }
 
-Preview.propTypes = {
-  proportion: PropTypes.number.isRequired
-};
-ResponsiveTabs.propTypes = {
-  deviceList: PropTypes.array,
-  action: PropTypes.func,
-  activeDevice: PropTypes.object,
-  styles: PropTypes.object,
-  parentWidth: PropTypes.number
-};
 ReactSpecimen.propTypes = {
   catalog: catalogShape.isRequired,
   children: PropTypes.oneOfType([PropTypes.element, PropTypes.string]).isRequired,
